refactor: move header components into its subfolder
